pyproject.toml
[project]
name = "semantic-search-mcp"
version = "0.1.0"
description = ""
authors = [
    {name = "Marat Khabibullin",email = "Marat.Khabibullin@thomsonreuters.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "sentence-transformers (>=5.1.1,<6.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "typer[all] (>=0.19.2,<0.20.0)",
    "mcp (>=1.15.0,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "streamlit (>=1.28.0,<2.0.0)",
    "spacy (>=3.7.0,<4.0.0)",
    "langchain-text-splitters (>=0.3.0,<1.0.0)",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "boto3 (>=1.40.47,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "semantic_search_mcp", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=6.30.1,<7.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-cov (>=6.0.0,<7.0.0)"
]
